refactor(dashboards): update paths for dashboard ds imports
