+++ public/app/features/dashboard-scene/scene/DashboardDatasourceBehaviour.tsx
@@ -1,7 +1,7 @@
 import { Unsubscribable } from 'rxjs';
 
 import { SceneObjectBase, SceneObjectState, SceneQueryRunner, VizPanel } from '@grafana/scenes';
-import { SHARED_DASHBOARD_QUERY } from 'app/plugins/datasource/dashboard';
+import { SHARED_DASHBOARD_QUERY } from 'app/plugins/datasource/dashboard/constants';
 
 import {
   findVizPanelByKey,
