+++ public/app/features/dashboard/components/DashboardRow/DashboardRow.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 
 import { createTheme } from '@grafana/data';
 import { selectors } from '@grafana/e2e-selectors';
-import { SHARED_DASHBOARD_QUERY } from 'app/plugins/datasource/dashboard/types';
+import { SHARED_DASHBOARD_QUERY } from 'app/plugins/datasource/dashboard/constants';
 
 import { PanelModel } from '../../state/PanelModel';
