+++ public/app/plugins/datasource/dashboard/constants.ts
@@ -1,0 +1,2 @@
+export const SHARED_DASHBOARD_QUERY = '-- Dashboard --';
+export const DASHBOARD_DATASOURCE_PLUGIN_ID = 'dashboard';
