+++ public/app/plugins/datasource/dashboard/types.ts
@@ -1,8 +1,5 @@
 import { DataFrame, DataQuery, DataQueryError, DataTopic } from '@grafana/data';
 
-export const SHARED_DASHBOARD_QUERY = '-- Dashboard --';
-export const DASHBOARD_DATASOURCE_PLUGIN_ID = 'dashboard';
-
 export interface DashboardQuery extends DataQuery {
   panelId?: number;
   withTransforms?: boolean;
